// src/models/RoutesTypes.ts
enum RoutesTypes {
    ADMIN = "/admin",
    HOME = "/home",
    LOGIN = "/login",
    REGISTER = "/register",
    CITATION = "/citations",
    PUNCHLINE = "/punchline",
    ADD_CITATION = "/ajouter-citation",
    LIST_CITATION = "/list-citations",
    UPDATE_CITATION = "/modifier-citation",
    DELETE_CITATION = "/supprimer-citation",
    ADD_PUNCHLINE = "/ajouter-punchline",
    LIST_PUNCHLINE = "/list-punchline",
    UPDATE_PUNCHLINE = "/modifier-punchline",
    DELETE_PUNCHLINE = "/supprimer-punchline",
}

export default RoutesTypes;